Refactor the CLI to be separate commands (#61)
* Refactor the CLI to be separate commands

This refactor aims to make the commands more organized and easier to
work it.  The idea behind this is that we can add new commands and
separate the logic for all of them in their own individual modules.

Currently, this is a rewrite of what we had in __main__.py, so we have a
`query`, `history` and `record` command. In the future, we may have more
or less.

* Fix specfile build

* Group test parameters with parametrize