[RSPEED-279] Add Command Line Assistant Daemon (#58)
* Refactor config module

This patch introduces a small refactor to the config module to split it
and handle the different schemas in a different module. The idea behind
this is to keep the modules unclogged and simple.

* Refactor CLI to have separate commands

This patch will introduce a refactor/split of our CLI module. The
intention here is to make sure that each (sub)command will be treated
independently and have their own logic.

Currently, it is being divided into two categories:
* history
* query

The history subcommand will handle everything that is related to the
user history of conversation.

The query subcommand will handle anything that is related to user
queries. This is also the default subcommand.

* [RSPEED-299] Initial setup for clad (#40)

* Initial setup for clad

This patch introduces an initial setup for clad (Command Line Assistant
Daemon).

* Fix unit-tests workflow

* Update CONTRIBUTING.md

* Drop python3-gobject-base from specfile

* Remove version lock from pyproject.toml

* Refactor the build specfile (#56)

In this patch, we refactored how we did the build for the CLA project.
This was needed, because of a couple of reasons:

- We were not including the submodules under `command_line_assistant` source package

- The binaries we were shipping were not working properly with root
-- Any time that we tried to execute the binary through `$ c`, it tried
to load from `lib64` instead of `lib`

To achieve those modifications, we created a new folder called `bin`,
where it has `c` and `clad`.

We also modified the pyproject.toml to include the submodules using the
Find Directive.

* Add authentication with identity certificate (#59)

This was not tested yet, but the code is all here.

* Add the python3-colorama dependency in specfile (#64)

* Change DBUS service name (#63)

* Improve rendering libary (#65)