Devel-dbus (#285)
* Install D-Bus policy into datadir

sysconfdir (/etc) is typically reserved for local system administrator
overrides. The default D-Bus policy should be shipped in /usr/share/dbus
instead. System administrators can still override the policy by dropping
a file into /etc, but this allows D-Bus to fall back to the desired
policy.

* Include a very loose policy for development

This policy is very relaxed, intentionally, to simplify development. It
is not intended for use in production.