Adjust exit codes
main() exits with various values for each exception it catches:

- ValueError results in  return value of os.EX_DATAERR (64)
- DBusError results in return value of os.EX_UNAVAILABLE (69)
- RuntimeError results in return value of os.EX_SOFTWARE (70)
- KeyboardInterrupt no longer results in a non-zero exit.

Additionally:
- ChatCommandException has a class variable 'code' equal to 80
- ShellCommandException has a class variable 'code' equal to 81
- HistoryCommandException has a class variable 'code' equal to 82
- FeedbackCommandException has a class variable 'code' equal to 83

These exceptions are caught in their respective BaseCLICommand subclasses. If an exception is caught in the 'run' function, the code value is returned. This effectively means that if an error occurs deeper within the 'chat' subcommand, the exit code of the program will be 80. Likewise for 81, 82, and 83.