[RSPEED-1064] Increase question/context size to 32k (#372)
* Increase question/context size to 32k

The current 2k limit is not good enough and does not fit the customer
needs anymore. We need to bump that in order to provide more meaningful
answers to the users.

This 32k limit is to not overflow the LLM context window buffer.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update command_line_assistant/commands/chat.py

Co-authored-by: Sam Doran <github@samdoran.com>

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Sam Doran <github@samdoran.com>