[RSPEED-1554] Fix interactive mode regression (#389)
* Fix interactive mode regression

With the merge of #297 to main, we introduced a regression where the `Asking RHEL Lightspeed` message is not printed.

Since we went with a default of static message instead of one spinning, it does not make sense to never show that message to the user.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>