Catch OSError being thrown by submit method
When the submit method throws an error, catch OSErrors and inspect the message. If it contains "/etc/pki/consumer/cert.pem" (the default file path for the RHSM consumer certificate), assume the system is not registered with RHSM and raise a custom error string recommending the user connect the system using rhc.

Fixes: RSPEED-1762
Fixes: RHEL-102629
Fixes: #335